+++ ansible_kernel/kernel.py
@@ -37,7 +37,7 @@ TaskCompletionMessage = namedtuple('TaskCompletionMessage', ['task_num'])
 
 TASK_ARGS_MODULES = modules + task_args
 
-__version__ = '0.5.0'
+__version__ = '0.6.0'
 
 logger = logging.getLogger('ansible_kernel.kernel')